requirements.txt
flask
flask-sqlalchemy
flask-mail
openai
python-dotenv
psycopg2
